完成里程计

+++ src/driver/src/driver-node/util.cpp
@@ -6,7 +6,7 @@ namespace driverNodeUtil
 {
 
 // 转速转换比例，执行速度调整比例
-float GetCoef(float item)
+int GetCoef(double item)
 {
     if (item > 128)
     {
@@ -18,7 +18,7 @@ float GetCoef(float item)
     }
     else
     {
-        return item + 128;
+        return static_cast<int>(item + 128);
     }
 };
